refactor(deep_causality_cfd): split three oversized flow modules by seam
Three files had grown past the point where the file boundary told you
anything: corridor.rs at 1226 LOC, study/mod.rs at 1087, and their two
test files at 1363 and 1072. Each is now a directory module split along
a seam the code already had, with no file over 350 LOC.

src/types/flow/corridor.rs -> corridor/ (1226 -> 6 files, max 348)
  The module docs already numbered the stages [2]/[3] through [6]; the
  split follows those numbers exactly. regime.rs classifies, trajectory_nav.rs
  navigates, branch.rs scores counterfactual branches and steers the aero
  force, envelope.rs carries the verified limits, gate.rs enforces them, and
  mod.rs keeps the docs, the re-exports, and the shared scalar/vector helpers.

src/types/flow/study/mod.rs -> 5 siblings (1087 -> 213, max 324)
  Grouped by which DSL path each verb belongs to: sweep_phase.rs (pointwise
  and march-every-case), event_fork.rs (fork/branch/continue_for), refine.rs
  (the refinement round), origin_fork.rs (baseline/alternate/ensemble plus the
  audit paths), and reduce.rs (reduction, recording, gating, terminal verdict).
  mod.rs keeps the phase-state vocabulary every sibling impls onto.

tests/types/flow/{corridor,compressible_march_run}_tests.rs -> directories
  Mirroring the source seams: corridor/ splits into regime/navigation/lift/
  gate/burn, and compressible_march_run/ into carrier/fork/study/audit/
  forcing/imprint. Shared fixtures live in each mod.rs, so the world builders
  and step context have one definition.

No behavior change, and no API change: the public items of both source
directories diff identically against HEAD, lib.rs is untouched, and the
pub(crate) count is unchanged at 5. No visibility had to be widened —
child modules can reach an ancestor's private items, so the phase-state
and stage fields stayed private.

One build-system consequence worth knowing: the Bazel types_flow target
globs "types/flow/**/*_tests.rs", which cargo's module-tree walk does not
depend on. The first naming (carrier.rs, fork.rs) built green under cargo
and clippy while failing under Bazel with "file not found for module". The
split test files therefore follow the existing *_tests.rs convention, and
the two new mod.rs files are registered explicitly rather than widening
the glob.
